2 simple activities at 5 october 2023
Design a mobile application that collects user information including their name, contact number, and email ID. Upon clicking the 'Submit' button, the application should display this information. How would you outline the user interface and functionality of this application? 

Design a mobile application for a university that allows users to input student information including the student's name, select a course from a dropdown list, and enter marks for three subjects: Java, DBMS, and Python. When the user clicks on the 'Submit' button, the application should calculate and display the percentage of the student based on the provided marks. Outline the user interface and functionality of this application.